research: 시장 폭 정식 검증 — 지수에는 넣지 않는다 (기각)
업종지수 44개가 다 모여 3분할로 정식 검증했다. 답은 '넣지 마라' 였다.

  기준 10개        검증 -0.650   테스트 -0.636
  + breadth 2개    (검증에서 채택)  테스트 -0.619   ← 나빠짐

검증 구간에서 채택된 건 '200일선 위 업종비율'과 '소형-대형' 두 개인데,
한 번도 보지 않은 2021년 이후에서 오히려 성능이 떨어졌다.

후보들이 구간별로 정반대로 움직인 것이 이유다.
                     검증(2016-20)  테스트(2021~)
  200일선 위 업종비율      +0.064       +0.391
  소형-대형              -0.380       +0.030
  분산 매도일             -0.278       -0.023
  지수-폭 괴리            +0.001       -0.411
검증에서 강했던 것이 테스트에서 사라지고, 약했던 것이 강해졌다.
표본 8회 문제에서 흔한 불안정성이고, 검증 구간을 보고 골랐다면
정확히 반대로 골랐을 것이다.

단독으로도 단조성이 약하다(고점 근처 5분위: -6.8% → -4.7%, 중간이 뒤섞임).

47개 지수를 받으려고 KRX 차단까지 맞아가며 붙였지만, 절차의 답이
'기각'이면 기각한다. 데이터는 남고 화면에는 참고 지표로 띄운다 —
현재 200일선 위 업종 61%, 소형-대형 -7.9%.

test_breadth_does_not_improve_index 가 이 결론을 고정한다.
나중에 개선되는 쪽으로 깨지면 '좋아졌다'가 아니라 '절차가 지켜졌나'를
먼저 확인해야 한다.

테스트 231개 통과.

+++ src/research/breadth.py
@@ -94,7 +94,8 @@ def divergence(index_close: pd.Series, breadth: pd.Series,
     from src.indicators.technical import pct_rank
 
     px = index_close.dropna()
-    bd = breadth.dropna().reindex(px.index).ffill()
+    # astype 없이 ffill 하면 bool/object 로 남아 pandas 가 downcast 경고를 낸다
+    bd = breadth.dropna().astype("float64").reindex(px.index).ffill()
     px_pr = pct_rank(px.rename("px"), window)
     bd_pr = pct_rank(bd.rename("bd"), window)
     return (px_pr - bd_pr).dropna().rename(f"divergence_{window}")
